requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
requests>=2.31.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
